add buffer interface and impl
